Conflicts:
	application/modules/hardwareoptimization/views/scripts/report/dealer/optimization/docx/00-render.phtml
	application/modules/hardwareoptimization/views/scripts/report/dealer/optimization/docx/02_costanalysis.phtml
	application/modules/hardwareoptimization/views/scripts/report/dealer/optimization/docx/03_devicesummary.phtml
	application/modules/hardwareoptimization/views/scripts/report/dealer/optimization/docx/04_replacements.phtml
	application/modules/hardwareoptimization/views/scripts/report/dealer/optimization/docx/05_acceptable.phtml
	application/modules/hardwareoptimization/views/scripts/report/dealer/optimization/docx/06_excess.phtml
	application/modules/hardwareoptimization/views/scripts/report/dealer/optimization/docx/07_retired.phtml
	application/modules/hardwareoptimization/views/scripts/report/dealer/optimization/docx/08_flagged.phtml
	application/modules/hardwareoptimization/views/scripts/report/dealer/optimization/docx/09_leased.phtml
	application/modules/hardwareoptimization/views/scripts/report/dealer/optimization/docx/10_excluded.phtml

+++ application/modules/proposalgen/forms/MasterDevice.php
@@ -203,7 +203,7 @@ class Proposalgen_Form_MasterDevice extends EasyBib_Form
          * Launch Date
          */
         $minYear    = 1950;
-        $maxYear    = date('Y') + 2;
+        $maxYear    = ((int)date('Y')) + 2;
         $launchDate = new ZendX_JQuery_Form_Element_DatePicker('launch_date');
 
         $launchDate->setJQueryParam('dateFormat', 'yy-mm-dd')
